services/auth/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy[asyncio]
asyncpg
pydantic-settings
python-jose[cryptography]
cryptography
httpx
redis
pgvector
PyJWT
